qr-hint-code/requirements.txt
jpype1
z3-solver
colorclass
terminaltables
matplotlib
numpy
pandas
boolean_parser
networkx
